# backend/application/src/main/resources/api/v1/specs/openapi.yaml
openapi: 3.0.3
info:
  title: Strategy & Planning API
  version: 0.1.0
  description: |
    Strategy & Planning toolkit API. Provides endpoints for the Case Study Builder,
    Category Analysis Builder, and RFP Outline Generator tools. Authentication via Clerk SSO (Bearer JWT).
servers:
  - url: /api/v1
    description: V1 base path (DOCUMENTATION ONLY — the openapi-generator-maven-plugin
      does NOT prepend this to generated Spring routes. Path keys below carry the
      `/api/v1/` prefix verbatim. See canonical-openapi-rules.md → "/api/v1 prefix").
security:
  - bearerAuth: [ ]
...
paths:
...
  /api/v1/rfp-outlines:
    get:
      tags: [ RfpOutlines ]
      operationId: listRfpOutlines
      summary: List all RFP outlines for the authenticated user
      description: Returns all RFP outlines created by the authenticated user, ordered by creation date descending.
      security: [ { bearerAuth: [ ] } ]
      responses:
        "200":
          description: List of RFP outline summaries for the authenticated user.
          content:
            application/json:
              schema:
                type: array
                description: Ordered list of RFP outline summaries.
                items:
                  $ref: "#/components/schemas/RfpOutlineSummaryV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while listing RFP outlines.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"

    post:
      tags: [ RfpOutlines ]
      operationId: createRfpOutline
      summary: Create a new RFP outline and export it to Google Docs
      description: |
        Creates a new RFP outline with the provided reviewed fields and immediately generates a Google Doc
        by copying the tokenized master template into the user's Drive and filling all template
        tokens. Requires the user to be signed in with Google (via Clerk) with Drive and Docs
        access. On success the record is persisted with status SAVED and the doc URL is returned
        in docUrl.
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        required: true
        description: RFP outline fields. Title is required; all other fields are optional pre-fill.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRfpOutlineRequestV1"
      responses:
        "201":
          description: RFP outline created and its Google Doc generated successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RfpOutlineV1"
        "400":
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppValidationExceptionResponseV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while creating the RFP outline.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "502":
          description: >-
            Google doc generation failed — the user is not connected to Google, has not granted
            Drive & Docs access, or a Google API call failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
  /api/v1/rfp-outlines/drafts:
    post:
      tags: [ RfpOutlines ]
      operationId: draftRfpOutline
      summary: AI-draft a POV and deck outline from an uploaded RFP
      description: |
        Accepts up to 10 source documents (PDF, DOCX, DOC, PPTX, PPT, XLSX, XLS, or CSV), extracts
        their readable content, and uses the AI drafting engine — together with optional
        supplementary notes — to draft a succinct POV (challenge, opportunity, solution, outcome)
        and a slide-by-slide response deck outline. The result is meant to be reviewed and edited
        by the user before submission.
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        required: true
        description: Source documents to draft from (1-10 files), plus optional supplementary notes.
        content:
          multipart/form-data:
            schema:
              type: object
              description: Multipart payload carrying the uploaded RFP documents and optional notes.
              required: [ files ]
              properties:
                files:
                  type: array
                  description: Uploaded RFP documents (PDF, DOCX, DOC, PPTX, PPT, XLSX, XLS, or CSV). Maximum 10 files.
                  minItems: 1
                  maxItems: 10
                  items:
                    type: string
                    format: binary
                    description: One uploaded RFP document.
                notes:
                  type: string
                  description: Optional free-text supplementary context alongside the uploaded RFP.
      responses:
        "200":
          description: Drafted POV narrative and deck outline for review.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RfpOutlineDraftV1"
        "400":
          description: Request validation failed (no files, more than 10 files, or an unsupported/unreadable file).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "502":
          description: AI drafting engine not configured or the drafting call failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while drafting the RFP outline.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
  /api/v1/rfp-outlines/{id}:
    get:
      tags: [ RfpOutlines ]
      operationId: getRfpOutline
      summary: Get an RFP outline by ID
      description: Returns the full detail of an RFP outline belonging to the authenticated user.
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the RFP outline.
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: RFP outline detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RfpOutlineV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "404":
          description: RFP outline not found or does not belong to the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while retrieving the RFP outline.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
...
  /api/v1/rnd-requests:
    get:
      tags: [ RndRequests ]
      operationId: listRndRequests
      summary: List all RnD requests for the authenticated user
      description: Returns all triaged RnD requests created by the authenticated user, ordered by creation date descending.
      security: [ { bearerAuth: [ ] } ]
      responses:
        "200":
          description: List of RnD request summaries for the authenticated user.
          content:
            application/json:
              schema:
                type: array
                description: Ordered list of RnD request summaries.
                items:
                  $ref: "#/components/schemas/RndRequestSummaryV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while listing RnD requests.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"

    post:
      tags: [ RndRequests ]
      operationId: createRndRequest
      summary: Create and triage a new RnD request
      description: |
        Creates a new RnD request and triages it immediately based on the deal size ("the buy").
        When the buy meets or exceeds the escalation threshold (default $100,000) the request is
        marked ESCALATE_TO_RND and a drafted RnD submission summary is returned. Otherwise it is
        marked WORKAROUND and a positive, relay-ready workaround response is drafted — the tool
        never produces a flat "no". Capability search across Slite/Asana is included in the
        capability summary when those integrations are connected; until then the summary reports
        they are not connected.
      security: [ { bearerAuth: [ ] } ]
      requestBody:
        required: true
        description: RnD request fields. Title and buy amount are required.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRndRequestRequestV1"
      responses:
        "201":
          description: RnD request created and triaged successfully.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RndRequestV1"
        "400":
          description: Request validation failed.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppValidationExceptionResponseV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while creating the RnD request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
  /api/v1/rnd-requests/{id}:
    get:
      tags: [ RndRequests ]
      operationId: getRndRequest
      summary: Get an RnD request by ID
      description: Returns the full detail of a triaged RnD request belonging to the authenticated user.
      security: [ { bearerAuth: [ ] } ]
      parameters:
        - name: id
          in: path
          required: true
          description: The unique identifier of the RnD request.
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: RnD request detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RndRequestV1"
        "401":
          description: The authorization token not provided or expired.
        "403":
          description: The authorization token not correct or user not authorized.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "404":
          description: RnD request not found or does not belong to the authenticated user.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
        "500":
          description: Exception occurred while retrieving the RnD request.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppApiExceptionResponseV1"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    GenerationStatusV1:
      type: string
      description: |
        Lifecycle status of a generated document.
        DRAFT — saved but not yet submitted for generation.
        SUBMITTED — submitted for Google Slides generation.
        GENERATED — Google Slides deck created in the user's Drive.
      enum: [ DRAFT, SUBMITTED, GENERATED ]
...
    RfpOutlineSummaryV1:
      type: object
      additionalProperties: false
      description: Lightweight summary of an RFP outline for list views.
      required: [ id, title, status, createdAt ]
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the RFP outline.
        title:
          type: string
          description: Title of the RFP outline.
        clientName:
          type: string
          nullable: true
          description: Name of the client this RFP outline responds to.
        industry:
          type: string
          nullable: true
          description: Industry or sector of the client.
        status:
          description: Lifecycle state of the RFP outline's document generation.
          allOf:
            - $ref: "#/components/schemas/GenerationStatusV1"
        createdAt:
          type: string
          format: date-time
          description: UTC timestamp when the RFP outline was created.
    RfpOutlineV1:
      type: object
      additionalProperties: false
      description: Full detail of an RFP outline including all fields and generation metadata.
      required: [ id, title, status, createdAt ]
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the RFP outline.
        title:
          type: string
          description: Title of the RFP outline.
        clientName:
          type: string
          nullable: true
          description: Name of the client this RFP outline responds to.
        industry:
          type: string
          nullable: true
          description: Industry or sector of the client.
        challenge:
          type: string
          nullable: true
          description: The client's core challenge or need.
        opportunity:
          type: string
          nullable: true
          description: The differentiator or angle that makes AI Digital's approach compelling for this client.
        solution:
          type: string
          nullable: true
          description: The proposed strategic approach at a high level.
        outcome:
          type: string
          nullable: true
          description: What success looks like.
        deckOutline:
          type: string
          nullable: true
          description: Slide-by-slide breakdown of how the response deck should be structured.
        supplementaryNotes:
          type: string
          nullable: true
          description: Free-text context supplied alongside the uploaded RFP, if provided.
        sourceFileName:
          type: string
          nullable: true
          description: Comma-separated names of the uploaded source files used for auto-drafting, if provided.
        docUrl:
          type: string
          nullable: true
          description: URL of the generated Google Doc in the user's Drive, when created.
        status:
          description: Lifecycle state of the RFP outline's document generation.
          allOf:
            - $ref: "#/components/schemas/GenerationStatusV1"
        createdAt:
          type: string
          format: date-time
          description: UTC timestamp when the RFP outline was created.
    CreateRfpOutlineRequestV1:
      type: object
      additionalProperties: false
      description: Request payload for creating a new RFP outline. Title is required; all other fields are optional pre-fill.
      required: [ title ]
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 255
          description: Title of the RFP outline.
        clientName:
          type: string
          maxLength: 255
          nullable: true
          description: Name of the client this RFP outline responds to.
        industry:
          type: string
          maxLength: 255
          nullable: true
          description: Industry or sector of the client.
        challenge:
          type: string
          maxLength: 5000
          nullable: true
          description: The client's core challenge or need.
        opportunity:
          type: string
          maxLength: 5000
          nullable: true
          description: The differentiator or angle that makes AI Digital's approach compelling for this client.
        solution:
          type: string
          maxLength: 5000
          nullable: true
          description: The proposed strategic approach at a high level.
        outcome:
          type: string
          maxLength: 5000
          nullable: true
          description: What success looks like.
        deckOutline:
          type: string
          maxLength: 20000
          nullable: true
          description: Slide-by-slide breakdown of how the response deck should be structured.
        supplementaryNotes:
          type: string
          maxLength: 20000
          nullable: true
          description: Free-text context supplied alongside the uploaded RFP, if provided.
        sourceFileName:
          type: string
          maxLength: 2000
          nullable: true
          description: Comma-separated names of the uploaded source files used for auto-drafting, if provided.
    RfpOutlineDraftV1:
      type: object
      additionalProperties: false
      description: |
        AI-drafted RFP outline fields extracted from the uploaded RFP source documents (plus any
        supplementary notes). Every field is nullable — the model only fills what the input supports.
      required: [ sourceFileNames ]
      properties:
        title:
          type: string
          nullable: true
          description: Drafted title of the RFP outline.
        clientName:
          type: string
          nullable: true
          description: Drafted client name.
        industry:
          type: string
          nullable: true
          description: Drafted industry or vertical.
        challenge:
          type: string
          nullable: true
          description: Drafted core challenge or need.
        opportunity:
          type: string
          nullable: true
          description: Drafted differentiator or angle for this client.
        solution:
          type: string
          nullable: true
          description: Drafted proposed strategic approach.
        outcome:
          type: string
          nullable: true
          description: Drafted description of what success looks like.
        deckOutline:
          type: string
          nullable: true
          description: Drafted slide-by-slide response deck outline.
        sourceFileNames:
          type: string
          description: Comma-separated names of the uploaded source files the draft was produced from.
...
    RndDecisionV1:
      type: string
      description: |
        Triage decision for an RnD request.
        ESCALATE_TO_RND — the buy meets or exceeds the escalation threshold; a drafted RnD submission summary is provided.
        WORKAROUND — the buy is below the threshold; a positive, relay-ready workaround response is provided instead of a "no".
      enum: [ ESCALATE_TO_RND, WORKAROUND ]
    RndRequestSummaryV1:
      type: object
      additionalProperties: false
      description: Lightweight summary of a triaged RnD request for list views.
      required: [ id, title, buyAmount, decision, status, createdAt ]
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the RnD request.
        title:
          type: string
          description: Short title of the requested feature or product capability.
        requesterTeam:
          type: string
          nullable: true
          description: Growth team or requester the ask came from.
        buyAmount:
          type: number
          description: Deal size ("the buy") in USD used for triage.
        decision:
          description: Triage outcome for the request, derived from the deal size.
          allOf:
            - $ref: "#/components/schemas/RndDecisionV1"
        status:
          description: Lifecycle state of the request's triage run.
          allOf:
            - $ref: "#/components/schemas/GenerationStatusV1"
        createdAt:
          type: string
          format: date-time
          description: UTC timestamp when the RnD request was created.
    RndRequestV1:
      type: object
      additionalProperties: false
      description: Full detail of a triaged RnD request including the drafted relay-ready response.
      required: [ id, title, buyAmount, decision, status, createdAt ]
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier of the RnD request.
        title:
          type: string
          description: Short title of the requested feature or product capability.
        requesterTeam:
          type: string
          nullable: true
          description: Growth team or requester the ask came from.
        requestDetails:
          type: string
          nullable: true
          description: Full description of the feature/product request from the growth team.
        capabilityNotes:
          type: string
          nullable: true
          description: Manual notes about existing capabilities or possible workarounds supplied by the triager.
        buyAmount:
          type: number
          description: Deal size ("the buy") in USD used for triage.
        decision:
          description: Triage outcome for the request, derived from the deal size.
          allOf:
            - $ref: "#/components/schemas/RndDecisionV1"
        capabilitySummary:
          type: string
          nullable: true
          description: Result of the capability search across connected knowledge sources (Slite, Asana). Reports "not connected" when integrations are not configured.
        responseDraft:
          type: string
          nullable: true
          description: Drafted output — an RnD submission summary when escalated, otherwise a positive relay-ready workaround response.
        status:
          description: Lifecycle state of the request's triage run.
          allOf:
            - $ref: "#/components/schemas/GenerationStatusV1"
        createdAt:
          type: string
          format: date-time
          description: UTC timestamp when the RnD request was created.
    CreateRndRequestRequestV1:
      type: object
      additionalProperties: false
      description: Request payload for creating and triaging a new RnD request. Title and buy amount are required.
      required: [ title, buyAmount ]
      properties:
        title:
          type: string
          maxLength: 255
          minLength: 1
          description: Short title of the requested feature or product capability. Must not be blank.
        requesterTeam:
          type: string
          maxLength: 255
          nullable: true
          description: Growth team or requester the ask came from.
        requestDetails:
          type: string
          maxLength: 5000
          nullable: true
          description: Full description of the feature/product request.
        capabilityNotes:
          type: string
          maxLength: 5000
          nullable: true
          description: Manual notes about existing capabilities or possible workarounds.
        buyAmount:
          type: number
          description: Deal size ("the buy") in USD. Must be zero or greater.
          minimum: 0
    AppValidationExceptionResponseV1:
      type: object
      additionalProperties: false
      description: Representation of validation exception occurred in REST API format.
      required: [ errors, timestamp, correlationId ]
      properties:
        errors:
          description: The list of validation errors occurred.
          type: array
          items:
            $ref: "#/components/schemas/FieldToErrorResponseV1"
        timestamp:
          type: string
          format: date-time
          description: The timestamp exception occurred.
        correlationId:
          type: string
          description: The log stream UUID.
    FieldToErrorResponseV1:
      type: object
      additionalProperties: false
      description: Representation of validation error bound to a field or semantic request part.
      required: [ code, field, error ]
      properties:
        code: { type: string, description: "The code of validation error occurred." }
        field: { type: string, description: "The field or semantic request part not passed validation." }
        error: { type: string, description: "The validation error occurred." }
    AppApiExceptionResponseV1:
      type: object
      additionalProperties: false
      description: Representation of exception occurred in REST API format.
      required: [ code, message, timestamp, correlationId ]
      properties:
        code: { type: string, description: "The code of exception occurred." }
        message: { type: string, description: "The message of exception occurred." }
        timestamp: { type: string, format: date-time, description: "The timestamp exception occurred." }
        correlationId: { type: string, description: "The log stream UUID." }